Fix duplicate `audit` subcommand call
When I wrote this dockerfile, I made some weird choices -- such as to
implicitly add args to the `cargo-audit` container. This confused me
months later when the `action.yaml` wasn't calling the `audit` subcommand
so I added it -- but this broke containers.

Rather than revert, I'm removing the subcommand from being in the dockerfile
since it doesn't make a lot of sense to be here.